+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Versioning].
 
 ## [Unreleased]
 
+- Re-enable the Action from a Pull Request comment. ([#129])
 - Fix default commit message description. ([#130])
 
 ## [0.3.1] - 2020-03-12
@@ -80,4 +81,5 @@ Versioning].
 [#112]: https://github.com/ericcornelissen/svgo-action/pull/112
 [#122]: https://github.com/ericcornelissen/svgo-action/pull/122
 [#124]: https://github.com/ericcornelissen/svgo-action/pull/124
+[#129]: https://github.com/ericcornelissen/svgo-action/pull/129
 [#130]: https://github.com/ericcornelissen/svgo-action/pull/130